requirements.txt
asgiref==3.5.2
Django==4.1
django-filter==22.1
django-cors-headers==3.7.0
djangorestframework==3.13.1
PyJWT==2.4.0
pytz==2022.2.1
pillow==9.2.0
sqlparse==0.4.2
tzdata==2022.2
